refactor: #130 - DataLayer 관련 리펙토링
- Data Layer는 UIkit을 import하지 않아야 함으로 Data -> CVPixelBuffer로 변경하는 로직으로 수정
- UIkit -> CoreGraphics, CoreVideo, ImageIO로 수정
