+++ BestWish/Data/Network/Manager/CoreML/CoreMLManager.swift
@@ -5,28 +5,25 @@
 //  Created by Quarang on 6/23/25.
 //
 
-import Foundation
 import CoreML
-import UIKit
 
 /// CoreML 데이터 추출 매니저
 final class CoreMLManager {
     /// 라벨 데이터 추출
-    func fetchCoreMLLabelData(image: UIImage) throws -> [LabelData] {
+    func fetchCoreMLLabelData(imageData: Data) throws -> [LabelDataEntity] {
         do {
             let model = try BestWidhClassfication()
-            guard let buffer = image.toCVPixelBuffer() else { return [] }
+            guard let buffer = imageData.toCVPixelBuffer() else { return [] }
             
             let output = try model.prediction(image: buffer)
             let labels = output.targetProbability
                 .sorted(by: { $0.value > $1.value })
                 .map {
-                    LabelData(label: $0.key, probability: Int($0.value * 100))
+                    LabelDataEntity(label: $0.key, probability: Int($0.value * 100))
                 }
             return labels
         } catch {
             throw CoreMLError.modelLoadingFailed(error)
         }
-        
     }
 }
