+++ BestWish/Data/Util/Data+Util.swift
@@ -1,0 +1,58 @@
+//
+//  Data+CVPixelBuffer.swift
+//  BestWish
+//
+//  Created by Quarang on 6/17/25.
+//
+
+import CoreGraphics
+import CoreVideo
+import ImageIO
+
+// MARK: - CoreML에서 이미지 형식 변화
+extension Data {
+    /// CoreML 이미지 입력 형식 변환
+    func toCVPixelBuffer() -> CVPixelBuffer? {
+        guard let imageSource = CGImageSourceCreateWithData(self as CFData, nil),
+              let cgImage = CGImageSourceCreateImageAtIndex(imageSource, 0, nil) else {
+            return nil
+        }
+        let width = cgImage.width
+        let height = cgImage.height
+
+        let attrs: CFDictionary = [
+            kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue!,
+            kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue!
+        ] as CFDictionary
+
+        var pixelBuffer: CVPixelBuffer?
+        let status = CVPixelBufferCreate(kCFAllocatorDefault,
+                                         width,
+                                         height,
+                                         kCVPixelFormatType_32BGRA,
+                                         attrs,
+                                         &pixelBuffer)
+
+        guard let buffer = pixelBuffer, status == kCVReturnSuccess else { return nil }
+
+        CVPixelBufferLockBaseAddress(buffer, [])
+
+        let colorSpace = CGColorSpaceCreateDeviceRGB()
+        guard let context = CGContext(data: CVPixelBufferGetBaseAddress(buffer),
+                                      width: width,
+                                      height: height,
+                                      bitsPerComponent: 8,
+                                      bytesPerRow: CVPixelBufferGetBytesPerRow(buffer),
+                                      space: colorSpace,
+                                      bitmapInfo: CGImageAlphaInfo.premultipliedFirst.rawValue)
+        else {
+            CVPixelBufferUnlockBaseAddress(buffer, [])
+            return nil
+        }
+
+        context.draw(cgImage, in: CGRect(x: 0, y: 0, width: width, height: height))
+        CVPixelBufferUnlockBaseAddress(buffer, [])
+
+        return buffer
+    }
+}
